Downgrade GMP to 6.2.1 due missing symbols
Appears when combined with Zig for cross-compilation, v6.3.0 of GMP
reports missing symbols when targetting x86_64 and aarch64 Linux musl.

This is not present when cross-compiling to macOS versions or when
compiling natively (using `cc` from hydrofoil-crystal), which is even
weirder considering that hydrofoil-crystal 1.11 and 1.12 already uses
v3.19 of Alpine as base and that was already using GMP v6.3.0.

Downgrading until I can investigate this further.